implemented remote SDK for product service

+++ product/spec/model.go
@@ -1,10 +1,10 @@
 package spec
 
 type Product struct {
-	ID          int    `json:"id"`
-	Name        string `json:"name"`
-	Price       int    `json:"price"`
-	CategoryID  int    `json:"-"`
-	Category    string `json:"category"`
-	AvlQuantity int    `json:"availableQuantity"`
+	ID          int     `json:"id"`
+	Name        string  `json:"name"`
+	Price       float64 `json:"price"`
+	CategoryID  int     `json:"-"`
+	Category    string  `json:"category"`
+	AvlQuantity int     `json:"availableQuantity"`
 }
